include ccn-lite-riot.h in ccnl-fwd.c to make cache_strategy_cache() available
This is almost definitely not the correct way to do it as ccnl-fwd
shouldn't depend on ccn-lite-riot.h. The root problem is that all of the
caching logic is only defined in ccn-lite-riot.{c,h}, when it should
probably be defined somewhere in the CCNL core instead. So this is just
a dirty fix for now.